Added default_config module.

+++ local-prepare.sh
@@ -1,8 +1,5 @@
 #!/usr/bin/env bash
 
-# Install composer dependencies inside of APIBus
-docker-compose run api_bus composer install
-
 # Install composer inside of Gifts Backend.
 docker-compose run be_gifts composer install
 
@@ -18,17 +15,32 @@ cp ./backend-api-bus/src/config/local.default.php ./backend-api-bus/src/config/l
 # Boot the dev environments.
 docker-compose up -d
 
-# Prepares certificates for Gifts backend.
-docker-compose run be_gifts openssl genrsa -out ./certificates/private.key 2048
-docker-compose run be_gifts openssl rsa -in ./certificates/private.key -pubout > backend-gifts/certificates/public.key
+read -p "Would you like to reinstall the platform? IMPORTANT: it will remove existing data in your databases. [Y/n]" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  # Prepares certificates for Gifts backend.
+  docker-compose run be_gifts openssl genrsa -out ./certificates/private.key 2048
+  docker-compose run be_gifts openssl rsa -in ./certificates/private.key -pubout > backend-gifts/certificates/public.key
 
-# Use generated keys for Donations backend.
-cp ./backend-gifts/certificates/public.key ./backend-donations/certificates/public.key
-cp ./backend-gifts/certificates/private.key ./backend-donations/certificates/private.key
+  # Use generated keys for Donations backend.
+  cp ./backend-gifts/certificates/public.key ./backend-donations/certificates/public.key
+  cp ./backend-gifts/certificates/private.key ./backend-donations/certificates/private.key
 
-cp ./backend-gifts/web/sites/default/example.settings.local.php ./backend-gifts/web/sites/default/settings.local.php
-cp ./backend-donations/web/sites/default/example.settings.local.php ./backend-donations/web/sites/default/settings.local.php
+  cp ./backend-gifts/web/sites/default/example.settings.local.php ./backend-gifts/web/sites/default/settings.local.php
+  cp ./backend-donations/web/sites/default/example.settings.local.php ./backend-donations/web/sites/default/settings.local.php
 
-# Run site installation step
-docker-compose run be_gifts drush site-install config_installer --root='./web' --yes
-docker-compose run be_donations drush site-install config_installer --root='./web' --yes
+  # Run site installation.
+  docker-compose run be_gifts drush site-install config_installer --root='./web' --yes
+  # To create a demo content.
+  docker-compose run be_gifts drush en flc_demo --root='./web' --yes
+  # Disable unnecessary modules.
+  docker-compose run be_donations drush pmu flc_demo default_content better_normalizers hal --root='./web' --yes
+
+  # Run site installation.
+  docker-compose run be_donations drush site-install config_installer --root='./web' --yes
+  # To create a demo content.
+  docker-compose run be_donations drush en flc_demo --root='./web' --yes
+  # Disable unnecessary modules.
+  docker-compose run be_donations drush pmu flc_demo default_content better_normalizers hal --root='./web' --yes
+fi
